GCD_SPI
SystemVerilog code that generates hardware to calculate the greatest common denominator, outputs the final value using SPI protocol and makes this output signal available to be observed in an oscilloscope.  It includes a number of files including data paths, finite state machines, a clock divider, a button debouncer, a test bench, etc.