Genomatic (aka DNA sequencer)
SystemVerilog code to generate hardware that develops an FPGA-based design that reads data from an embedded memory and determines the frequency of occurrence of substrings defined in a second memory. The frequency of each substring is recorded in individual registers which can be accessed by switches on a Xilinx Zybo board.